Mutualize move error handler
